check in directive directory with factored scroll listener
