+++ app/js/app.js
@@ -15,8 +15,13 @@ require('../vendor/css/navisworks-a360');
 require('../vendor/js/three.min');
 var autodeskCredentials = require('./lib/autodesk.js');
 
+// directives
+require('./directives/scroll-listener');
+
 // main application
-var buildingApp = angular.module("BuildingApp", ["ngRoute", "ngSanitize", "rzModule"]);
+var buildingApp = angular.module("BuildingApp",
+  ["ngRoute", "ngSanitize", "rzModule", "ScrollListener"]
+);
 
 /***
 *
@@ -77,8 +82,6 @@ buildingApp.config([
 
 }]);
 
-
-
 /***
 *
 * Add a run block to update scroll position on request of
@@ -118,53 +121,6 @@ buildingApp.directive('backgroundImage', function(){
     });
   };
 });
-
-
-/***
-*
-* Directive to listen for scroll events
-*
-***/
-
-buildingApp.directive('scrollListener', function () {
-  return {
-    restrict: 'AC',
-
-    // give the directive access to the controller function
-    // specified in the html directive, so that when the
-    // user scrolls, the directive can broadcast the new
-    // scroll position to the subscribing controller
-    scope: {
-      setScrollPosition: '='
-    },
-
-    // bind a scroll event listener to the selected element
-    // and broadcast the scroll position to the subscribing
-    // controller function
-    link: function (scope, element, attrs) {
-      element.bind('scroll', function () {
-
-        // pass the scroll position to the subscriber's setScrollPosition method
-        var scrollPosition = element[0].scrollTop;
-
-        // if the scroll position > 0, fade in articles after the 1st
-        // then permanently set their opacity to 1, else fade them out
-        if (scrollPosition > 0) {
-          // set timeout so that fade in completes before we absolutely set opacity
-          $(".major-section-spacer *").removeClass("fade-out-article");
-          $(".major-section-spacer *").addClass("fade-in-article");
-        } else {
-          // set timeout so that fade out completes before we absolutely set opacity
-          $(".major-section-spacer *").removeClass("fade-in-article");
-          $(".major-section-spacer *").addClass("fade-out-article");
-        }
-
-        scope.setScrollPosition(scrollPosition);
-      });
-    }
-  };
-});
-
 
 /***
 *
